+++ spec/active_record/pg_enum_spec.rb
@@ -3,7 +3,6 @@ require "spec_helper"
 RSpec.describe ActiveRecord::PGEnum do
   class TestTable < ActiveRecord::Base
     self.table_name = "test_table"
-    include ActiveRecord::PGEnum::Helper
   end
 
   let(:column) { TestTable.columns.detect { |col| col.name == "foo" } }
@@ -22,15 +21,13 @@ RSpec.describe ActiveRecord::PGEnum do
     expect { TestTable.create!(foo: "quux") }.to raise_error(ActiveRecord::StatementInvalid)
   end
 
-  describe ActiveRecord::PGEnum::Helper do
+  describe "ActiveRecord::PGEnum()" do
     subject { TestTable }
-
-    it { is_expected.to respond_to :pg_enum }
 
     it "converts arrays of values into hashes for ::enum" do
       expect(TestTable).to receive(:enum).with(foo_type: { bar: "bar", baz: "baz" }).and_call_original
 
-      TestTable.pg_enum foo_type: %i[bar baz]
+      TestTable.include ActiveRecord.PGEnum(foo_type: %i[bar baz])
 
       expect(TestTable).to respond_to :foo_types
       expect(TestTable.foo_types).to match({ "bar" => "bar", "baz" => "baz" })